apps/base/serializers/tenant.py
#!/usr/bin/env python3
# -*- coding=utf-8 -*-

""" 
@version: v1.0 
@author: cm   
@time: 02/03/2018 10:12 AM 
"""

__all__ = ['TenantSerializer']

from rest_framework import serializers
from ..models import Tenant


class TenantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tenant
        fields = '__all__'
