requirements.txt
Django==2.0.2
django-debug-toolbar==1.9.1
django-extensions==2.0.0
django-filter==1.1.0
django-js-asset==1.0.0
django-mptt==0.9.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Pillow==5.0.0
psycopg2-binary==2.7.4
PyJWT==1.5.3
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
typing==3.6.4
